requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Session==0.4.0
Faker==18.13.0
python-dateutil==2.8.2
reportlab==4.0.4
openpyxl==3.1.2
XlsxWriter==3.1.2
WeasyPrint==60.1 